+++ android/src/main/java/com/capacitorjs/community/plugins/bluetoothle/BluetoothLe.kt
@@ -3,13 +3,18 @@ package com.capacitorjs.community.plugins.bluetoothle
 import android.Manifest
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
+import android.bluetooth.BluetoothGattCharacteristic
 import android.bluetooth.BluetoothManager
 import android.bluetooth.le.ScanFilter
 import android.bluetooth.le.ScanResult
 import android.bluetooth.le.ScanSettings
+import android.content.BroadcastReceiver
 import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
 import android.content.pm.PackageManager
 import android.os.ParcelUuid
+import android.util.Log
 import com.getcapacitor.*
 import com.getcapacitor.annotation.CapacitorPlugin
 import com.getcapacitor.annotation.Permission
@@ -39,6 +44,7 @@ class BluetoothLe : Plugin() {
     }
 
     private var bluetoothAdapter: BluetoothAdapter? = null
+    private var stateReceiver: BroadcastReceiver? = null
     private var deviceMap = HashMap<String, Device>()
     private var deviceScanner: DeviceScanner? = null
 
@@ -64,7 +70,7 @@ class BluetoothLe : Plugin() {
         // Use this check to determine whether BLE is supported on the device. Then
         // you can selectively disable BLE-related features.
         if (!activity.packageManager.hasSystemFeature(PackageManager.FEATURE_BLUETOOTH_LE)) {
-            call.reject("BLE is not supported")
+            call.reject("BLE is not supported.")
             return
         }
 
@@ -72,12 +78,61 @@ class BluetoothLe : Plugin() {
         bluetoothAdapter = (activity.getSystemService(Context.BLUETOOTH_SERVICE)
                 as BluetoothManager).adapter
 
-        // Ensures Bluetooth is available on the device and it is enabled.
-        if (bluetoothAdapter == null || bluetoothAdapter?.isEnabled != true) {
-            bluetoothAdapter = null
-            call.reject("BLE is not enabled.")
+        if (bluetoothAdapter == null) {
+            call.reject("BLE is not available.")
             return
         }
+        call.resolve()
+    }
+
+    @PluginMethod
+    fun getEnabled(call: PluginCall) {
+        assertBluetoothAdapter(call) ?: return
+        val enabled = bluetoothAdapter?.isEnabled == true
+        val result = JSObject()
+        result.put("value", enabled)
+        call.resolve(result)
+    }
+
+    @PluginMethod
+    fun startEnabledNotifications(call: PluginCall) {
+        assertBluetoothAdapter(call) ?: return
+        createStateReceiver()
+
+        try {
+            val intentFilter = IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED);
+            context.registerReceiver(stateReceiver, intentFilter);
+        } catch (e: Error) {
+            call.reject("Error")
+            return
+        }
+        call.resolve()
+    }
+
+    private fun createStateReceiver() {
+        if (stateReceiver == null) {
+            stateReceiver = object : BroadcastReceiver() {
+                override fun onReceive(context: Context, intent: Intent) {
+                    val action = intent.action
+                    if (action == BluetoothAdapter.ACTION_STATE_CHANGED) {
+                        val state = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE,
+                                BluetoothAdapter.ERROR)
+                        val enabled = state == BluetoothAdapter.STATE_ON
+                        val result = JSObject()
+                        result.put("value", enabled)
+                        notifyListeners("onEnabledChanged", result)
+                    }
+                }
+            }
+        }
+    }
+
+    @PluginMethod
+    fun stopEnabledNotifications(call: PluginCall) {
+        if (stateReceiver != null) {
+            context.unregisterReceiver(stateReceiver)
+        }
+        stateReceiver = null
         call.resolve()
     }
 
@@ -178,7 +233,13 @@ class BluetoothLe : Plugin() {
 
         val device: Device
         try {
-            device = Device(activity.applicationContext, bluetoothAdapter!!, deviceId)
+            device = Device(
+                    activity.applicationContext,
+                    bluetoothAdapter!!,
+                    deviceId
+            ) { ->
+                onDisconnect(deviceId)
+            }
         } catch (e: IllegalArgumentException) {
             call.reject("Invalid deviceId")
             return
@@ -193,6 +254,10 @@ class BluetoothLe : Plugin() {
                 }
             }
         }
+    }
+
+    private fun onDisconnect(deviceId: String) {
+        notifyListeners("disconnected|${deviceId}", null)
     }
 
     @PluginMethod
@@ -237,7 +302,29 @@ class BluetoothLe : Plugin() {
             call.reject("Value required.")
             return
         }
-        device.write(characteristic.first, characteristic.second, value) { response ->
+        val writeType = BluetoothGattCharacteristic.WRITE_TYPE_DEFAULT
+        device.write(characteristic.first, characteristic.second, value, writeType) { response ->
+            run {
+                if (response.success) {
+                    call.resolve()
+                } else {
+                    call.reject(response.value)
+                }
+            }
+        }
+    }
+
+    @PluginMethod
+    fun writeWithoutResponse(call: PluginCall) {
+        val device = getDevice(call) ?: return
+        val characteristic = getCharacteristic(call) ?: return
+        val value = call.getString("value", null)
+        if (value == null) {
+            call.reject("Value required.")
+            return
+        }
+        val writeType = BluetoothGattCharacteristic.WRITE_TYPE_NO_RESPONSE
+        device.write(characteristic.first, characteristic.second, value, writeType) { response ->
             run {
                 if (response.success) {
                     call.resolve()