+++ android/src/main/java/com/capacitorjs/community/plugins/bluetoothle/Device.kt
@@ -17,6 +17,7 @@ class Device(
         private val context: Context,
         private val bluetoothAdapter: BluetoothAdapter,
         private val address: String,
+        private val onDisconnect: () -> Unit
 ) {
     companion object {
         private val TAG = Device::class.java.simpleName
@@ -51,6 +52,7 @@ class Device(
                 }
             } else if (newState == BluetoothProfile.STATE_DISCONNECTED) {
                 connectionState = STATE_DISCONNECTED
+                onDisconnect()
                 Log.d(TAG, "Disconnected from GATT server.")
                 resolve("disconnect", "Disconnected.")
             }
@@ -184,7 +186,13 @@ class Device(
         setTimeout(key, "Read timeout.")
     }
 
-    fun write(serviceUUID: UUID, characteristicUUID: UUID, value: String, callback: (CallbackResponse) -> Unit) {
+    fun write(
+            serviceUUID: UUID,
+            characteristicUUID: UUID,
+            value: String,
+            writeType: Int,
+            callback: (CallbackResponse) -> Unit
+    ) {
         val key = "write|$serviceUUID|$characteristicUUID"
         callbackMap[key] = callback
         val service = bluetoothGatt?.getService(serviceUUID)
@@ -192,7 +200,7 @@ class Device(
         if (characteristic != null) {
             val bytes = stringToBytes(value)
             characteristic.value = bytes
-            characteristic.writeType = WRITE_TYPE_DEFAULT
+            characteristic.writeType = writeType
             val result = bluetoothGatt?.writeCharacteristic(characteristic)
             if (result != true) {
                 reject(key, "Writing characteristic failed.")