+++ ios/Plugin/DeviceManager.swift
@@ -3,6 +3,7 @@ import CoreBluetooth
 
 class DeviceManager: NSObject, CBCentralManagerDelegate {
     typealias Callback = (_ success: Bool, _ message: String) -> Void
+    typealias StateReceiver = (_ enabled: Bool) -> Void
     typealias ScanResultCallback = (_ device: Device, _ advertisementData: [String: Any], _ rssi: NSNumber) -> Void
 
     private var centralManager: CBCentralManager!
@@ -10,6 +11,7 @@ class DeviceManager: NSObject, CBCentralManagerDelegate {
     private var displayStrings: [String: String]!
     private var callbackMap = [String: Callback]()
     private var scanResultCallback: ScanResultCallback?
+    private var stateReceiver: StateReceiver?
     private var timeoutMap = [String: DispatchWorkItem]()
     private var stopScanWorkItem: DispatchWorkItem?
     private var alertController: UIAlertController?
@@ -29,17 +31,43 @@ class DeviceManager: NSObject, CBCentralManagerDelegate {
 
     // initialize
     func centralManagerDidUpdateState(_ central: CBCentralManager) {
-        let key = "initialize"
+        let initializeKey = "initialize"
         switch central.state {
         case .poweredOn:
-            self.resolve(key, "BLE powered on")
+            self.resolve(initializeKey, "BLE powered on")
+            self.emitState(enabled: true)
         case .poweredOff:
-            central.stopScan()
-            self.reject(key, "BLE powered off")
+            self.stopScan()
+            self.resolve(initializeKey, "BLE powered off")
+            self.emitState(enabled: false)
+        case .resetting:
+            self.emitState(enabled: false)
+        case .unauthorized:
+            self.emitState(enabled: false)
         case .unsupported:
-            self.reject(key, "BLE unsupported")
+            self.reject(initializeKey, "BLE unsupported")
+            self.emitState(enabled: false)
+        case .unknown:
+            self.emitState(enabled: false)
         default: break
         }
+    }
+
+    func getEnabled() -> Bool {
+        return self.centralManager.state == CBManagerState.poweredOn
+    }
+
+    func registerStateReceiver( _ stateReceiver: @escaping StateReceiver) {
+        self.stateReceiver = stateReceiver
+    }
+
+    func unregisterStateReceiver() {
+        self.stateReceiver = nil
+    }
+
+    func emitState(enabled: Bool) {
+        guard let stateReceiver = self.stateReceiver else { return }
+        stateReceiver(enabled)
     }
 
     func startScanning(
@@ -72,7 +100,7 @@ class DeviceManager: NSObject, CBCentralManagerDelegate {
                 }
                 DispatchQueue.main.asyncAfter(deadline: .now() + scanDuration!, execute: self.stopScanWorkItem!)
             }
-            self.centralManager.scanForPeripherals(withServices: serviceUUIDs, options: [CBCentralManagerScanOptionAllowDuplicatesKey: false])
+            self.centralManager.scanForPeripherals(withServices: serviceUUIDs, options: [CBCentralManagerScanOptionAllowDuplicatesKey: allowDuplicates])
 
             if shouldShowDeviceList == false {
                 self.resolve("startScanning", "Scan started.")
@@ -174,8 +202,13 @@ class DeviceManager: NSObject, CBCentralManagerDelegate {
         self.reject(key, "Failed to connect.")
     }
 
+    func setOnDisconnected(_ device: Device, _ callback: @escaping Callback) {
+        let key = "onDisconnected|\(device.getId())"
+        self.callbackMap[key] = callback
+    }
+
     func disconnect(_ device: Device, _ callback: @escaping Callback) {
-        let key = "disconnect|\(device.getPeripheral().identifier.uuidString)"
+        let key = "disconnect|\(device.getId())"
         self.callbackMap[key] = callback
         print("Disconnecting from peripheral", device.getPeripheral())
         self.centralManager.cancelPeripheralConnection(device.getPeripheral())
@@ -185,7 +218,10 @@ class DeviceManager: NSObject, CBCentralManagerDelegate {
     // didDisconnectPeripheral
     func centralManager(_ central: CBCentralManager, didDisconnectPeripheral peripheral: CBPeripheral, error: Error?) {
         let key = "disconnect|\(peripheral.identifier.uuidString)"
+        let keyOnDisconnected = "onDisconnected|\(peripheral.identifier.uuidString)"
+        self.resolve(keyOnDisconnected, "Disconnected.")
         if error != nil {
+            print(error!.localizedDescription)
             self.reject(key, error!.localizedDescription)
             return
         }